Rename package to conform with site name
