+++ spring-thymeleaf/src/main/java/com/baeldung/thymeleaf/config/WebApp.java
@@ -1,0 +1,36 @@
+package com.baeldung.thymeleaf.config;
+
+import javax.servlet.ServletRegistration.Dynamic;
+
+import org.springframework.web.servlet.support.AbstractAnnotationConfigDispatcherServletInitializer;
+
+/**
+ * Java configuration file that is used for web application initialization
+ */
+public class WebApp extends AbstractAnnotationConfigDispatcherServletInitializer {
+
+	public WebApp() {
+		super();
+	}
+
+	@Override
+	protected Class<?>[] getRootConfigClasses() {
+		return null;
+	}
+
+	@Override
+	protected Class<?>[] getServletConfigClasses() {
+		return new Class<?>[] { WebMVCConfig.class };
+	}
+
+	@Override
+	protected String[] getServletMappings() {
+		return new String[] { "/" };
+	}
+
+	@Override
+	protected void customizeRegistration(final Dynamic registration) {
+		super.customizeRegistration(registration);
+	}
+
+}
